+++ app/src/main/java/com/techcourse/ManualHandlerMapping.java
@@ -1,5 +1,6 @@
 package com.techcourse;
 
+import com.techcourse.controller.regacy.ForwardController;
 import com.techcourse.controller.regacy.LoginController;
 import com.techcourse.controller.regacy.LoginViewController;
 import com.techcourse.controller.regacy.LogoutController;
@@ -10,7 +11,6 @@ import java.util.Optional;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import webmvc.org.springframework.web.servlet.mvc.asis.Controller;
-import webmvc.org.springframework.web.servlet.mvc.asis.ForwardController;
 import webmvc.org.springframework.web.servlet.mvc.tobe.HandlerMapping;
 
 public class ManualHandlerMapping implements HandlerMapping {
